backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ask-my-docs-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",

    # ML / RAG
    "sentence-transformers>=3.3.0",
    "torch>=2.5.0",

    # File parsing
    "pymupdf>=1.25.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.5",

    # Config
    "pydantic-settings>=2.6.0",

    # HTTP client (for Ollama)
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "anyio>=4.7.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    # Eval
    "ragas>=0.2.0",
    "datasets>=3.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
